Split core from main
- separate core MiniLogicEva-functions from the main-class
